문제 추가 Language_Coder/문자열2/형성평가

+++ Language_Coder/문자열1/자가진단/BasicString108.cpp
@@ -22,7 +22,7 @@ void BasicString108::Code()
 
 	int count{ 1 };
 
-	for (int i = 0; i < str.size(); i++)
+	for (size_t i = 0; i < str.size(); i++)
 	{
 		if (str[i] == ' ')
 		{
