+++ Language_Coder/문자열1/형성평가/FormativeString103.cpp
@@ -18,7 +18,7 @@ void FormativeString103::Code()
 
 	std::cin >> str;
 
-	for (int i = 0; i < str.size(); i++)
+	for (size_t i = 0; i < str.size(); i++)
 	{
 		if (isalpha(str[i]))
 		{
